proj: add datasets and scripts
Two scripts are present. One script checks the seeds file,
and gives it as input to second script. The second script
uses that as seed and randomly selects rows from the dataset.

Second script takes a factor. We multiply the factor with
the number of attributes in the csv file to identify number
of rows we need to take out of original csv.

Signed-off-by: Mohan Yelugoti <ymdatta.work@gmail.com>